SBS-68: fix statistics

+++ Sobes/Libraries/Types/Sources/Types/Sources/InterviewQuestion.swift
@@ -5,10 +5,10 @@ public struct InterviewQuestion: Identifiable {
     public let id: Int
     public let questionType: Professions
     public let text: String
-    public let result: Double?
+    public let result: Int?
     public var messages: [InterviewMessage]
 
-    public init(id: Int, questionType: Professions, text: String, result: Double? = nil, messages: [InterviewMessage] = []) {
+    public init(id: Int, questionType: Professions, text: String, result: Int? = nil, messages: [InterviewMessage] = []) {
         self.id = id
         self.questionType = questionType
         self.text = text
